Refactoring

+++ users/khaykingleb/scripts/conda.sh
@@ -1,9 +1,14 @@
 #! /bin/bash
 
 install_conda_macos_silicon() {
-    curl -O https://repo.anaconda.com/archive/Anaconda3-2024.10-1-MacOSX-arm64.sh \
-        && sh Anaconda3-2024.10-1-MacOSX-arm64.sh -b -p ~/anaconda3 \
-        && rm Anaconda3-2024.10-1-MacOSX-arm64.sh
+    if [ ! -d "$HOME/anaconda3" ]; then
+        echo "Installing Anaconda..."
+        curl -O https://repo.anaconda.com/archive/Anaconda3-2024.10-1-MacOSX-arm64.sh \
+            && sh Anaconda3-2024.10-1-MacOSX-arm64.sh -b -p ~/anaconda3 \
+            && rm Anaconda3-2024.10-1-MacOSX-arm64.sh
+    else
+        echo "Anaconda is already installed, skipping installation..."
+    fi
 }
 
 "$@"
